api/requirements.txt
flask==3.0.0
flask-cors==4.0.0
paho-mqtt==1.6.1
python-dotenv==1.0.0
requests==2.31.0

# Testing dependencies
pytest==7.4.0
pytest-cov==4.1.0
